+++ wpull/app_test.py
@@ -135,6 +135,7 @@ class TestApp(GoodAppTestCase):
             '--no-strong-robots',
             '--restrict-file-names', 'windows,lower',
             '--quota', '10m',
+            '--max-filename-length', '100',
         ])
         with cd_tempdir():
             builder = Builder(args)
@@ -420,6 +421,52 @@ class TestApp(GoodAppTestCase):
         self.assertIn('somewhereelse.invalid', resolver.hosts_touched)
 
     @tornado.testing.gen_test(timeout=DEFAULT_TIMEOUT)
+    def test_redirect_span_hosts_allow_linked(self):
+        arg_parser = AppArgumentParser()
+        args = arg_parser.parse_args([
+            self.get_url(
+                '/span_hosts?port={0}'.format(self.get_http_port())
+            ),
+            '--span-hosts-allow', 'linked-pages',
+            '--no-robots',
+            '--recursive',
+        ])
+        builder = Builder(args)
+        builder.factory.class_map['Resolver'] = MockDNSResolver
+
+        with cd_tempdir():
+            engine = builder.build()
+            exit_code = yield engine()
+        self.assertEqual(0, exit_code)
+        self.assertEqual(2, builder.factory['Statistics'].files)
+
+        resolver = builder.factory['Resolver']
+        self.assertIn('linked.test', resolver.hosts_touched)
+
+    @tornado.testing.gen_test(timeout=DEFAULT_TIMEOUT)
+    def test_redirect_span_hosts_page_requisites(self):
+        arg_parser = AppArgumentParser()
+        args = arg_parser.parse_args([
+            self.get_url(
+                '/span_hosts?port={0}'.format(self.get_http_port())
+            ),
+            '--span-hosts-allow', 'page-requisites',
+            '--no-robots',
+            '--page-requisites',
+        ])
+        builder = Builder(args)
+        builder.factory.class_map['Resolver'] = MockDNSResolver
+
+        with cd_tempdir():
+            engine = builder.build()
+            exit_code = yield engine()
+        self.assertEqual(0, exit_code)
+        self.assertEqual(2, builder.factory['Statistics'].files)
+
+        resolver = builder.factory['Resolver']
+        self.assertIn('pagereq.test', resolver.hosts_touched)
+
+    @tornado.testing.gen_test(timeout=DEFAULT_TIMEOUT)
     def test_strong_redirect(self):
         arg_parser = AppArgumentParser()
         args = arg_parser.parse_args([
@@ -577,6 +624,30 @@ class TestApp(GoodAppTestCase):
                 self.assertIn(b'Count: 10', data)
 
         self.assertEqual(0, exit_code)
+
+    @tornado.testing.gen_test(timeout=DEFAULT_TIMEOUT)
+    def test_sitemaps(self):
+        arg_parser = AppArgumentParser()
+        args = arg_parser.parse_args([
+            self.get_url('/'),
+            '--no-robots',
+            '--sitemaps',
+            '--recursive',
+        ])
+
+        with cd_tempdir():
+            builder = Builder(args)
+            engine = builder.build()
+            exit_code = yield engine()
+
+            print(list(os.walk('.')))
+            self.assertTrue(os.path.exists(
+                'localhost:{0}/static/my_file.txt'.format(
+                    self.get_http_port())
+            ))
+
+        self.assertEqual(0, exit_code)
+        self.assertGreaterEqual(4, builder.factory['Statistics'].files)
 
 
 class TestAppBad(BadAppTestCase):