+++ wpull/writer.py
@@ -6,6 +6,7 @@ import base64
 import collections
 import email.utils
 import gettext
+import hashlib
 import http.client
 import itertools
 import logging
@@ -377,13 +378,14 @@ class PathNamer(BasePathNamer):
         cut (int): Number of leading directories to cut from the file path.
         protocol (bool): Include the URL scheme in the directory structure.
         hostname (bool): Include the hostname in the directory structure.
+        safe_filename_args (dict): Keyword arguments for `safe_filename`.
 
     See also: :func:`url_to_filename`, :func:`url_to_dir_path`,
     :func:`safe_filename`.
     '''
     def __init__(self, root, index='index.html', use_dir=False, cut=None,
     protocol=False, hostname=False, os_type='unix',
-    no_control=True, ascii_only=True, case=None):
+    no_control=True, ascii_only=True, case=None, max_filename_length=None):
         self._root = root
         self._index = index
         self._cut = cut
@@ -394,37 +396,55 @@ class PathNamer(BasePathNamer):
         self._no_control = no_control
         self._ascii_only = ascii_only
         self._case = case
+        self._max_filename_length = max_filename_length
 
     def get_filename(self, url_info):
         url = url_info.url
-        filename = url_to_filename(
-            url, self._index,
-            os_type=self._os_type, no_control=self._no_control,
-            ascii_only=self._ascii_only, case=self._case
-        )
+        alt_char = self._os_type == 'windows'
+        parts = []
 
         if self._use_dir:
-            dir_path = url_to_dir_path(
-                url, self._protocol, self._hostname,
-                os_type=self._os_type, no_control=self._no_control,
-                ascii_only=self._ascii_only, case=self._case
+            dir_parts = url_to_dir_parts(
+                url, self._protocol, self._hostname, alt_char=alt_char
             )
-            filename = os.path.join(dir_path, filename)
 
-        return filename
+            for dummy in range(self._cut or 0):
+                if dir_parts:
+                    del dir_parts[0]
+
+            parts.extend(dir_parts)
+
+        parts.append(url_to_filename(url, self._index, alt_char=alt_char))
+
+        parts = [
+            safe_filename(
+                part,
+                os_type=self._os_type, no_control=self._no_control,
+                ascii_only=self._ascii_only, case=self._case,
+                max_length=self._max_filename_length,
+            )
+            for part in parts
+        ]
+
+        return os.path.join(*parts)
 
 
-def url_to_filename(url, index='index.html', **safe_kwargs):
-    '''Return a safe filename from a URL.
+def url_to_filename(url, index='index.html', alt_char=False):
+    '''Return a filename from a URL.
 
     Args:
         url (str): The URL.
         index (str): If a filename could not be derived from the URL path,
             use index instead. For example, ``/images/`` will return
             ``index.html``.
-        safe_kwargs: Additional arguments to :func:`safe_filename`.
+        alt_char (bool): If True, the character for the query deliminator
+            will be ``@`` intead of ``?``.
 
-    This function does not include the directories.
+    This function does not include the directories and does not sanitize
+    the filename.
+
+    Returns:
+        str
     '''
     assert isinstance(url, str)
     url_split_result = urllib.parse.urlsplit(url)
@@ -434,10 +454,8 @@ def url_to_filename(url, index='index.html', **safe_kwargs):
     if not filename:
         filename = index
 
-    filename = safe_filename(filename, **safe_kwargs)
-
     if url_split_result.query:
-        if safe_kwargs.get('os_type') == 'windows':
+        if alt_char:
             query_delim = '@'
         else:
             query_delim = '?'
@@ -449,9 +467,9 @@ def url_to_filename(url, index='index.html', **safe_kwargs):
     return filename
 
 
-def url_to_dir_path(url, include_protocol=False, include_hostname=False,
-**safe_kwargs):
-    '''Return a safe directory path from a URL.
+def url_to_dir_parts(url, include_protocol=False, include_hostname=False,
+alt_char=False):
+    '''Return a list of directory parts from a URL.
 
     Args:
         url (str): The URL.
@@ -459,9 +477,14 @@ def url_to_dir_path(url, include_protocol=False, include_hostname=False,
             included.
         include_hostname (bool): If True, the hostname from the URL will be
             included.
-        safe_kwargs: Additional arguments to :func:`safe_filename`.
+        alt_char (bool): If True, the character for the port deliminator
+            will be ``+`` intead of ``:``.
 
-    This function does not include the filename.
+    This function does not include the filename and the paths are not
+    sanitized.
+
+    Returns:
+        list
     '''
     assert isinstance(url, str)
     url_split_result = urllib.parse.urlsplit(url)
@@ -475,7 +498,7 @@ def url_to_dir_path(url, include_protocol=False, include_hostname=False,
         hostname = url_split_result.hostname
 
         if url_split_result.port:
-            if safe_kwargs.get('os_type') == 'windows':
+            if alt_char:
                 port_delim = '+'
             else:
                 port_delim = ':'
@@ -493,11 +516,7 @@ def url_to_dir_path(url, include_protocol=False, include_hostname=False,
     if not url.endswith('/') and parts:
         parts.pop()
 
-    if not parts:
-        return ''
-
-    parts = [safe_filename(part, **safe_kwargs) for part in parts]
-    return os.path.join(*parts)
+    return parts
 
 
 class PercentEncoder(collections.defaultdict):
@@ -537,7 +556,7 @@ _encoder_cache = {}
 
 
 def safe_filename(filename, os_type='unix', no_control=True, ascii_only=True,
-case=None, encoding='utf8'):
+case=None, encoding='utf8', max_length=None):
     '''Return a safe filename or path part.
 
     Args:
@@ -549,6 +568,7 @@ case=None, encoding='utf8'):
         case (str): If ``lower``, lowercase the string. If ``upper``, uppercase
             the string.
         encoding (str): The character encoding.
+        max_length (int): The maximum length of the filename.
 
     This function assumes that `filename` has not already been percent-encoded.
 
@@ -581,6 +601,13 @@ case=None, encoding='utf8'):
             new_filename = '{0}{1:02X}'.format(
                 new_filename[:-1], new_filename[-1]
             )
+
+    if max_length and len(new_filename) > max_length:
+        hash_obj = hashlib.sha1(new_filename.encode(encoding))
+        new_length = max(0, max_length - 8)
+        new_filename = '{0}{1}'.format(
+            new_filename[:new_length], hash_obj.hexdigest()[:8]
+        )
 
     if case == 'lower':
         new_filename = new_filename.lower()